requirements.txt
Django==3.2.8
django-crispy-forms==1.13.0
Pillow==8.4.0
django-two-factor-auth==1.13.1
phonenumbers==8.12.36
django-utils-six==2.0.0